# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Essa é uma documentação simple do desafio proposto pela Phoebus para o carga de Engenheiro de Software Jr.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/rebelde/negociar": {
      "post": {
        "tags": [
          "negociar"
        ],
        "summary": "Negociar",
        "operationId": "negociar",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "rebeldeId1",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Código do rebelde 1"
          },
          {
            "name": "rebeldeId2",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Código do rebelde 2"
          },
          {
            "name": "negociarItens",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NegociarItensContext"
            },
            "description": "Itens a ser negociados"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Rebelde não encontrado"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Item": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "enum": [
            "ARMA",
            "MUNICAO",
            "AGUA",
            "COMIDA"
          ]
        },
        "quantidade": {
          "type": "number",
          "format": "integer"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "NegociarItensContext": {
      "type": "object",
      "properties": {
        "itensRebelde1": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        },
        "itensRebelde2": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    },
...
  },
...
}