# src/main/java/br/com/phoebus/desafiostarwars/model/NegociarItensContext.java
package br.com.phoebus.desafiostarwars.model;

import java.util.List;

import lombok.Getter;

@Getter
public class NegociarItensContext {

	private List<Item> itensRebelde1;
	
	private List<Item> itensRebelde2;
	
}